CloneToUninit: use a private specialization trait
and move implementation details into a submodule
